requirements.txt
httpx>=0.24.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
extruct>=0.15.0
w3lib>=2.1.0
python-dateutil>=2.8.0
pytz>=2023.3
pydantic>=2.0.0
jinja2>=3.1.0
feedparser>=6.0.0
icalendar>=5.0.0
rapidfuzz>=3.0.0
python-slugify>=8.0.0
tenacity>=8.2.0
openai>=1.0.0
mysql-connector-python>=9.0.0
